% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook}
\alias{codebook}
\title{create data frame with summary information about a data frame}
\usage{
codebook(dataframe)
}
\arguments{
\item{dataframe}{your data frame of choice}
}
\value{
a dataframe with length equal to number of columns in parent data frame
}
\description{
Creates a data frame about a dataframe.
usefull in codebook description
}

