% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_dat.R
\docType{package}
\name{summarizedat}
\alias{summarizedat}
\alias{summarizedat-package}
\title{Summarizedat Will Summarize Your Datafile In a Nice and Clean Way}
\description{
Once you cleaned and tested your tidy dataframe, you should create a summary of the
file. This package will act as a nicely formated str() and summary()
but will also perform as a codebook.
}

